+++ src/models/creatures/GrandeAraignee.java
@@ -48,7 +48,7 @@ public class GrandeAraignee extends Creature
 	 * @param nbPiecesDOr le nombre de pieces d'or de cette creature
 	 * @param vitesse vitesse de la creature
 	 */
-	public GrandeAraignee(long santeMax, int nbPiecesDOr, double vitesse)
+	public GrandeAraignee(double santeMax, int nbPiecesDOr, double vitesse)
 	{
 		this(0, 0, santeMax, nbPiecesDOr,vitesse);
 	}
@@ -66,7 +66,7 @@ public class GrandeAraignee extends Creature
 	 * @param nbPiecesDOr le nombre de pieces d'or de cette creature
 	 * @param vitesse vitesse de la creature
 	 */
-	public GrandeAraignee(int x, int y, long santeMax, int nbPiecesDOr, double vitesse)
+	public GrandeAraignee(int x, int y, double santeMax, int nbPiecesDOr, double vitesse)
 	{
 		super(x, y, 32, 32, santeMax, nbPiecesDOr, vitesse,
 		        Creature.TYPE_TERRIENNE, IMAGE, "Grande Araignee");
