+++ src/models/creatures/Paysan.java
@@ -53,7 +53,7 @@ public class Paysan extends Creature
 	 * @param nbPiecesDOr le nombre de pieces d'or de cette creature
 	 * @param vitesse vitesse de la creature
 	 */
-	public Paysan(long santeMax, int nbPiecesDOr, double vitesse)
+	public Paysan(double santeMax, int nbPiecesDOr, double vitesse)
 	{
 		this(0, 0, santeMax, nbPiecesDOr,vitesse);
 	}
@@ -72,7 +72,7 @@ public class Paysan extends Creature
 	 * @param nbPiecesDOr le nombre de pieces d'or de cette creature
 	 * @param vitesse vitesse de la creature
 	 */
-	public Paysan(int x, int y, long santeMax, int nbPiecesDOr, double vitesse)
+	public Paysan(int x, int y, double santeMax, int nbPiecesDOr, double vitesse)
 	{
 		super(x, y, 20, 20, santeMax, nbPiecesDOr, vitesse,
 		        Creature.TYPE_TERRIENNE, IMAGES[0], Constants.PAYSAN);
